+++ ICD.Connect.Routing.CrestronPro/DigitalMedia/DmMd6XN/IDmMd6XNAdapterSettings.cs
@@ -1,0 +1,6 @@
+﻿namespace ICD.Connect.Routing.CrestronPro.DigitalMedia.DmMd6XN
+{
+	public interface IDmMd6XNAdapterSettings : ICrestronSwitchAdapterSettings
+	{
+	}
+}