+++ ICD.Connect.Routing/Endpoints/AbstractSourceDestinationBaseSettings.cs
@@ -1,5 +1,7 @@
 ﻿using System.Collections.Generic;
+#if STANDARD
 using System.Linq;
+#endif
 using ICD.Common.Utils.Collections;
 using ICD.Common.Utils.Extensions;
 using ICD.Common.Utils.Xml;