+++ ICD.Connect.Routing/Extensions/CoreExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using ICD.Common.Utils.Extensions;
 using ICD.Connect.Routing.RoutingGraphs;
 using ICD.Connect.Settings.Core;
 
@@ -9,12 +10,31 @@ namespace ICD.Connect.Routing.Extensions
 	/// </summary>
 	public static class CoreExtensions
 	{
+		/// <summary>
+		/// Gets the first routing graph instance from the core.
+		/// </summary>
+		/// <param name="core"></param>
+		/// <returns></returns>
 		public static IRoutingGraph GetRoutingGraph(this ICore core)
 		{
 			if (core == null)
 				throw new ArgumentNullException("core");
 
 			return core.Originators.GetChild<IRoutingGraph>();
+		}
+
+		/// <summary>
+		/// Tries to get the first routing graph instance from the core.
+		/// </summary>
+		/// <param name="core"></param>
+		/// <param name="output"></param>
+		/// <returns></returns>
+		public static bool TryGetRoutingGraph(this ICore core, out IRoutingGraph output)
+		{
+			if (core == null)
+				throw new ArgumentNullException("core");
+
+			return core.Originators.GetChildren<IRoutingGraph>().TryFirst(out output);
 		}
 	}
 }